+++ RxFlow/FlowCoordinator.swift
@@ -97,7 +97,7 @@ public final class FlowCoordinator: NSObject {
             // the FlowContributor is not related to a new Flow but to a Presentable/Stepper
             // this new Stepper will contribute to the current Flow.
             .flatMap { [weak self] in
-                self?.steps(from: $0, within: flow, allowStepWhenDismissed: allowStepWhenDismissed) ?? Signal.empty()
+                self?.steps(from: $0) ?? Signal.empty()
             }
             .asObservable()
             .take(until: allowStepWhenDismissed ? .empty() : flow.rxDismissed.asObservable())
@@ -172,18 +172,19 @@ public final class FlowCoordinator: NSObject {
     /// retrieve Steps from the combination presentable/stepper
     ///
     /// - Parameter nextPresentableAndStepper: the combination presentable/stepper that will generate new Steps
-    /// - Parameter flow: the Flow in which the stepper emits new steps
     /// - Returns: the reactive sequence of Steps from the combination presentable/stepper
-    private func steps (from nextPresentableAndStepper: PresentableAndStepper,
-                        within flow: Flow,
-                        allowStepWhenDismissed: Bool = false) -> Signal<Step> {
+    private func steps(from nextPresentableAndStepper: PresentableAndStepper) -> Signal<Step> {
         var stepStream = nextPresentableAndStepper
             .stepper
             .steps
             .do(onSubscribed: { nextPresentableAndStepper.stepper.readyToEmitSteps() })
             .startWith(nextPresentableAndStepper.stepper.initialStep)
             .filter { !($0 is NoneStep) }
-            .take(until: allowStepWhenDismissed ? .empty() : nextPresentableAndStepper.presentable.rxDismissed.asObservable())
+            .take(
+                until: nextPresentableAndStepper.allowStepWhenDismissed
+                ? .empty()
+                : nextPresentableAndStepper.presentable.rxDismissed.asObservable()
+            )
 
         // by default we cannot accept steps from a presentable that is not visible
         if nextPresentableAndStepper.allowStepWhenNotPresented == false {