+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ** *Unreleased* **:
 - fix: flow is not released when embedded in UINavigationController
+- fix: presentable `allowStepWhenDismissed` now are not ignored and not overridden by parent flow
 - fix: `displayed` and `rxVisible` now do not assume UIViewController starts not visible
 
 ** Version 2.13.0 **: